requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
djangorestframework==3.13.1
idna==3.3
numpy==1.22.0
pandas==1.3.5
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
twilio==7.4.0
tzdata==2021.5
urllib3==1.26.7
